+++ components/browser/state/src/main/java/mozilla/components/browser/state/engine/EngineObserver.kt
@@ -21,7 +21,7 @@ import mozilla.components.browser.state.selector.findTabOrCustomTab
 import mozilla.components.browser.state.state.AppIntentState
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.LoadRequestState
-import mozilla.components.browser.state.state.SecurityInfoState
+import mozilla.components.browser.state.state.SecurityInfo
 import mozilla.components.browser.state.state.content.DownloadState
 import mozilla.components.browser.state.state.content.DownloadState.Status.INITIATED
 import mozilla.components.browser.state.state.content.FindResultState
@@ -166,7 +166,7 @@ internal class EngineObserver(
         dispatchAsync(
             ContentAction.UpdateSecurityInfoAction(
                 tabId,
-                SecurityInfoState(secure, host ?: "", issuer ?: "", certificate),
+                SecurityInfo.from(secure, host ?: "", issuer ?: "", certificate),
             ),
         )
     }