+++ components/feature/pwa/src/test/java/mozilla/components/feature/pwa/WebAppUseCasesTest.kt
@@ -7,7 +7,7 @@ package mozilla.components.feature.pwa
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import kotlinx.coroutines.test.runTest
 import mozilla.components.browser.state.state.BrowserState
-import mozilla.components.browser.state.state.SecurityInfoState
+import mozilla.components.browser.state.state.SecurityInfo
 import mozilla.components.browser.state.state.TabSessionState
 import mozilla.components.browser.state.state.createTab
 import mozilla.components.browser.state.store.BrowserStore
@@ -140,7 +140,7 @@ private fun createTestSession(
 
     return tab.copy(
         content = tab.content.copy(
-            securityInfo = SecurityInfoState(secure = secure),
+            securityInfo = SecurityInfo.from(secure),
             webAppManifest = manifest,
         ),
     )