+++ docs/changelog.md
@@ -5,6 +5,8 @@ permalink: /changelog/
 ---
 
 # 147.0 (In Development)
+* **browser-state**:
+  * ⚠️ **Breaking change**: New `Unknown` third state for the `SecurityStatus` of the connection used in a tab. The `secure` property has been migrated to `isSecure`. [Bug 2000617](https://bugzilla.mozilla.org/show_bug.cgi?id=2000617).
 * **lib-state**
     * ⚠️ **Breaking change**: Removed thread marshalling from Stores [Bug 1980348](https://bugzilla.mozilla.org/show_bug.cgi?id=1980348).
         * `Store.dispatch` is now invoked on the calling thread.